legacy/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pymongo==4.8.0
openai==1.58.1
anthropic==0.42.0
fastapi==0.115.6
uvicorn[standard]==0.34.0
typer==0.15.1
httpx==0.28.1
rich==13.9.4
feedparser==6.0.11
readability-lxml==0.8.1
lxml==5.3.0
pytest==8.3.4
